Fixed naming conventions for packages.
