--- src/main/java/dev/jacrispys/JavaBot/Audio/AudioPlayerSendHandler.java
@@ -1,39 +1,0 @@
-package dev.jacrispys.JavaBot.Audio;
-
-import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
-import com.sedmelluq.discord.lavaplayer.track.playback.MutableAudioFrame;
-import net.dv8tion.jda.api.audio.AudioSendHandler;
-import org.jetbrains.annotations.Nullable;
-
-import java.nio.Buffer;
-import java.nio.ByteBuffer;
-
-public class AudioPlayerSendHandler implements AudioSendHandler {
-    private AudioPlayer audioPlayer;
-    private final ByteBuffer buffer;
-    private final MutableAudioFrame frame;
-
-    public AudioPlayerSendHandler(AudioPlayer audioPlayer) {
-        this.audioPlayer = audioPlayer;
-        this.buffer = ByteBuffer.allocate(1024);
-        this.frame = new MutableAudioFrame();
-        this.frame.setBuffer(buffer);
-    }
-
-    @Override
-    public boolean canProvide() {
-        return audioPlayer.provide(frame);
-    }
-
-    @Nullable
-    @Override
-    public ByteBuffer provide20MsAudio() {
-        ((Buffer) buffer).flip();
-        return buffer;
-    }
-
-    @Override
-    public boolean isOpus() {
-        return true;
-    }
-}
