--- src/main/java/dev/jacrispys/JavaBot/Audio/objects/GuildBookmark.java
@@ -1,9 +1,0 @@
-package dev.jacrispys.JavaBot.Audio.objects;
-
-import dev.jacrispys.JavaBot.api.analytics.AudioGuildAnalytics;
-import net.dv8tion.jda.api.entities.Member;
-
-public class GuildBookmark {
-
-
-}
