--- src/main/java/dev/jacrispys/JavaBot/Audio/objects/GuildPlaylist.java
@@ -1,4 +1,0 @@
-package dev.jacrispys.JavaBot.Audio.objects;
-
-public class GuildPlaylist {
-}
