--- src/main/java/dev/jacrispys/JavaBot/Commands/RuntimeDebug/SlashDebugCommands.java
@@ -1,8 +1,0 @@
-package dev.jacrispys.JavaBot.Commands.RuntimeDebug;
-
-import net.dv8tion.jda.api.hooks.ListenerAdapter;
-
-public class SlashDebugCommands extends ListenerAdapter {
-
-
-}
