+++ src/main/java/dev/jacrispys/JavaBot/JavaBotMain.java
@@ -5,19 +5,19 @@ import com.github.topislavalinkplugins.topissourcemanagers.spotify.SpotifySource
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.player.DefaultAudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.source.AudioSourceManagers;
-import dev.jacrispys.JavaBot.Audio.AudioPlayerButtons;
-import dev.jacrispys.JavaBot.Audio.GenerateGenrePlaylist;
-import dev.jacrispys.JavaBot.Audio.InactivityTimer;
-import dev.jacrispys.JavaBot.Commands.Audio.GenericMusicCommands;
-import dev.jacrispys.JavaBot.Commands.Audio.SlashMusicCommands;
-import dev.jacrispys.JavaBot.Commands.*;
-import dev.jacrispys.JavaBot.Commands.PrivateMessageCommands.DefaultPrivateMessageResponse;
-import dev.jacrispys.JavaBot.Commands.RuntimeDebug.GenericDebugCommands;
-import dev.jacrispys.JavaBot.Events.BotStartup;
-import dev.jacrispys.JavaBot.Utils.GameSpyThread;
-import dev.jacrispys.JavaBot.Utils.MySQL.MySQLConnection;
-import dev.jacrispys.JavaBot.Utils.SecretData;
-import dev.jacrispys.JavaBot.Utils.SpotifyManager;
+import dev.jacrispys.JavaBot.audio.AudioPlayerButtons;
+import dev.jacrispys.JavaBot.audio.GenerateGenrePlaylist;
+import dev.jacrispys.JavaBot.audio.InactivityTimer;
+import dev.jacrispys.JavaBot.commands.audio.GenericMusicCommands;
+import dev.jacrispys.JavaBot.commands.audio.SlashMusicCommands;
+import dev.jacrispys.JavaBot.commands.*;
+import dev.jacrispys.JavaBot.commands.private_message.DefaultPrivateMessageResponse;
+import dev.jacrispys.JavaBot.commands.debug.GenericDebugCommands;
+import dev.jacrispys.JavaBot.events.BotStartup;
+import dev.jacrispys.JavaBot.utils.GameSpyThread;
+import dev.jacrispys.JavaBot.utils.mysql.MySQLConnection;
+import dev.jacrispys.JavaBot.utils.SecretData;
+import dev.jacrispys.JavaBot.utils.SpotifyManager;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.entities.Activity;
