+++ src/main/java/dev/jacrispys/JavaBot/api/analytics/AudioGuildAnalytics.java
@@ -1,13 +1,8 @@
 package dev.jacrispys.JavaBot.api.analytics;
 
-import dev.jacrispys.JavaBot.Audio.objects.GuildPlaylist;
-import dev.jacrispys.JavaBot.Audio.objects.GuildBookmark;
-import dev.jacrispys.JavaBot.api.analytics.objects.AudioUser;
-import dev.jacrispys.JavaBot.api.analytics.objects.PlayTime;
-import dev.jacrispys.JavaBot.api.analytics.objects.TrackStats;
-import net.dv8tion.jda.api.entities.UserSnowflake;
+import dev.jacrispys.JavaBot.audio.objects.GuildBookmark;
+import dev.jacrispys.JavaBot.audio.objects.GuildPlaylist;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public interface AudioGuildAnalytics extends AudioAnalytics, GeneralGuildAnalytics {
