+++ src/main/java/dev/jacrispys/JavaBot/audio/objects/GuildBookmark.java
@@ -1,0 +1,6 @@
+package dev.jacrispys.JavaBot.audio.objects;
+
+public class GuildBookmark {
+
+
+}
