+++ src/main/java/dev/jacrispys/JavaBot/audio/objects/GuildPlaylist.java
@@ -1,0 +1,4 @@
+package dev.jacrispys.JavaBot.audio.objects;
+
+public class GuildPlaylist {
+}
