+++ src/main/java/dev/jacrispys/JavaBot/commands/debug/SlashDebugCommands.java
@@ -1,0 +1,8 @@
+package dev.jacrispys.JavaBot.commands.debug;
+
+import net.dv8tion.jda.api.hooks.ListenerAdapter;
+
+public class SlashDebugCommands extends ListenerAdapter {
+
+
+}
